refactor: move benchmark util into shared utils
